requirements.txt
Flask
numpy
pandas 
SQLAlchemy 
psycopg2-binary
xlrd
openpyxl
pycountry